+++ pdf/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improve Table.fromTextArray()
 - Add curved LineDataSet Chart
 - Fix PdfColors.fromHex()
+- Update Barcode library to 1.9.0
 
 ## 1.7.1